backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "black>=25.1.0",
    "django>=5.2.5",
    "django-ckeditor>=6.7.3",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
]

[tool.black]
line-length = 100
target-version = ['py311']

[dependency-groups]
dev = [
    "black>=25.1.0",
]
